Refactor imports to use updated service paths
