+++ src/app/app.routes.ts
@@ -5,7 +5,7 @@ import { ActivityCreateComponent } from './components/activity/activity-create/a
 import { UserCreateComponent } from './components/user/user-create/user-create.component';
 import { UserEditComponent } from './components/user/user-edit/user-edit.component';
 import { HistoryComponent } from './components/history/history.component';
-import { AuthService } from './services/auth.service';
+import { AuthService } from './services/auth/auth.service';
 import { NotAuthenticatedComponent } from './components/not-authenticated/not-authenticated.component';
 
 export const routes: Routes = [
