+++ src/app/components/modal/sop-modal/sop-modal.component.ts
@@ -8,7 +8,7 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatButtonModule } from '@angular/material/button';
 import { FormsModule } from '@angular/forms';
-import { PdfService } from '../../../services/pdf.service';
+import { PdfService } from '../../../services/pdf/pdf.service';
 import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 
 @Component({
